Refactor user storage and auth methods for improved type safety and consistency
