+++ src/app/sign-in/sign-in.page.ts
@@ -36,7 +36,7 @@ export class SignInPage implements OnInit {
         this.authService.signin(form.value.email, form.value.password)
             .then(async (data) => {
                 loadingDialog.dismiss().catch(e => console.log('Could not dismiss loading dialog'));
-                await this.storeUserDataFromCredential(data);
+                await this.userStorageService.storeFromCredential(data);
                 
                 console.log('Navigating to shopping list..');
                 this.router.navigate(['/']).catch(e => console.log('Could not navigate'));
@@ -62,7 +62,7 @@ export class SignInPage implements OnInit {
         this.authService.signInWithGoogle()
             .then(async (data) => {
                 loadingDialog.dismiss().catch(e => console.log('Could not dismiss loading dialog'));
-                await this.storeUserDataFromCredential(data);
+                await this.userStorageService.storeFromCredential(data);
                 
                 console.log('Navigating to shopping list..');
                 this.router.navigate(['/']).catch(e => console.log('Could not navigate'));
@@ -78,16 +78,4 @@ export class SignInPage implements OnInit {
             });
     }
 
-    private async storeUserDataFromCredential(data: { user: any }): Promise<void> {
-        const user = data.user;
-        const userData: UserData = {
-            email: user.email || '',
-            uid: user.uid,
-            displayName: user.displayName || undefined,
-            photoURL: user.photoURL || undefined,
-            providerId: user.providerData[0]?.providerId || undefined,
-            lastLogin: new Date()
-        };
-        await this.userStorageService.storeUserData(userData);
-    }
 }
